pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forseti"
version = "0.1.0"
description = "⚖️ Forseti — LLM-Powered E2E Testing Service (เทพแห่งความยุติธรรม)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "MyHero Team" },
]
dependencies = [
    "click>=8.1",
    "playwright>=1.40",
    "google-genai>=1.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "pygithub>=2.1",
    "pydantic>=2.0",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
forseti = "forseti.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/forseti"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
